fire-detection/requirements.txt
# ml
numpy==1.22.0
tensorflow==2.8.0
keras==2.8.0
opencv-python==4.5.5.62
pillow==8.2.0
protobuf==3.19.0 # fix tf error
tf_explain==0.3.1

# web services
flask==2.1.2
flask-cors==3.0.10